Move ReaderWriterLockSlimExtensions into a separate file
