+++ Emby.AutoOrganize/Data/BaseSqliteRepository.cs
@@ -302,65 +302,4 @@ namespace Emby.AutoOrganize.Data
             connection.Execute("alter table " + table + " add column " + columnName + " " + type + " NULL");
         }
     }
-
-    public static class ReaderWriterLockSlimExtensions
-    {
-        public static IDisposable Read(this ReaderWriterLockSlim obj)
-        {
-            // if (BaseSqliteRepository.ThreadSafeMode > 0)
-            // {
-            //    return new DummyToken();
-            // }
-            return new WriteLockToken(obj);
-        }
-
-        public static IDisposable Write(this ReaderWriterLockSlim obj)
-        {
-            // if (BaseSqliteRepository.ThreadSafeMode > 0)
-            // {
-            //    return new DummyToken();
-            // }
-            return new WriteLockToken(obj);
-        }
-
-        private sealed class ReadLockToken : IDisposable
-        {
-            private ReaderWriterLockSlim _sync;
-
-            public ReadLockToken(ReaderWriterLockSlim sync)
-            {
-                _sync = sync;
-                sync.EnterReadLock();
-            }
-
-            public void Dispose()
-            {
-                if (_sync != null)
-                {
-                    _sync.ExitReadLock();
-                    _sync = null;
-                }
-            }
-        }
-
-        private sealed class WriteLockToken : IDisposable
-        {
-            private ReaderWriterLockSlim _sync;
-
-            public WriteLockToken(ReaderWriterLockSlim sync)
-            {
-                _sync = sync;
-                sync.EnterWriteLock();
-            }
-
-            public void Dispose()
-            {
-                if (_sync != null)
-                {
-                    _sync.ExitWriteLock();
-                    _sync = null;
-                }
-            }
-        }
-    }
 }
